+++ xla/python/ifrt/device_list.cc
@@ -15,20 +15,10 @@ limitations under the License.
 
 #include "xla/python/ifrt/device_list.h"
 
-#include <atomic>
-#include <cstdint>
-#include <initializer_list>
-#include <string>
-#include <utility>
 #include <vector>
 
-#include "absl/base/call_once.h"
-#include "absl/base/optimization.h"
 #include "absl/container/inlined_vector.h"
-#include "absl/hash/hash.h"
 #include "absl/status/statusor.h"
-#include "absl/strings/str_cat.h"
-#include "absl/strings/str_join.h"
 #include "absl/types/span.h"
 #include "xla/python/ifrt/client.h"
 #include "xla/python/ifrt/device.h"
@@ -40,7 +30,6 @@ namespace xla {
 namespace ifrt {
 
 char DeviceList::ID = 0;
-char BasicDeviceList::ID = 0;
 
 absl::StatusOr<tsl::RCReference<DeviceList>> DeviceList::FromProto(
     xla::ifrt::Client* client, const DeviceListProto& proto) {
@@ -61,69 +50,6 @@ DeviceListProto DeviceList::ToProto() const {
     proto.mutable_device_ids()->AddAlreadyReserved(device->Id().value());
   }
   return proto;
-}
-
-tsl::RCReference<DeviceList> BasicDeviceList::Create(Devices devices) {
-  return tsl::MakeRef<BasicDeviceList>(std::move(devices));
-}
-
-tsl::RCReference<DeviceList> BasicDeviceList::Create(
-    absl::Span<Device* const> devices) {
-  return Create(Devices(devices.begin(), devices.end()));
-}
-
-tsl::RCReference<DeviceList> BasicDeviceList::Create(
-    std::initializer_list<Device*> devices) {
-  return Create(Devices(devices.begin(), devices.end()));
-}
-
-BasicDeviceList::BasicDeviceList(Devices devices)
-    : devices_(std::move(devices)), hash_(kUnsetHash) {}
-
-DeviceList* BasicDeviceList::AddressableDeviceList() const {
-  absl::call_once(addressable_device_list_cache_.once_flag, [this] {
-    Devices addressable_devices;
-    for (Device* device : devices_) {
-      if (device->IsAddressable()) {
-        addressable_devices.push_back(device);
-      }
-    }
-    const bool already_fully_addressable =
-        addressable_devices.size() == devices_.size();
-    if (already_fully_addressable) {
-      // `device_list_holder` is intentionally unset. We skip storing a
-      // reference-counted copy in the holder to avoid creating a self cycle.
-      addressable_device_list_cache_.device_list =
-          const_cast<BasicDeviceList*>(this);
-    } else {
-      addressable_device_list_cache_.device_list_holder =
-          BasicDeviceList::Create(std::move(addressable_devices));
-      addressable_device_list_cache_.device_list =
-          addressable_device_list_cache_.device_list_holder.get();
-    }
-  });
-  return addressable_device_list_cache_.device_list;
-}
-
-uint64_t BasicDeviceList::hash() const {
-  uint64_t hash = hash_.load(std::memory_order_relaxed);
-  if (ABSL_PREDICT_FALSE(hash == kUnsetHash)) {
-    hash = absl::HashOf(devices());
-    if (ABSL_PREDICT_FALSE(hash == kUnsetHash)) {
-      ++hash;
-    }
-    hash_.store(hash, std::memory_order_relaxed);
-  }
-  return hash;
-}
-
-std::string BasicDeviceList::ToString() const {
-  return absl::StrCat("BasicDeviceList([",
-                      absl::StrJoin(devices_, ",",
-                                    [](std::string* out, Device* device) {
-                                      absl::StrAppend(out, device->ToString());
-                                    }),
-                      "])");
 }
 
 std::vector<DeviceId> GetDeviceIds(
